Continued on the parsing of the stages and their associated interactables.
Tidied up the interactable into the biome parsing. WHich was causing a crash. This is still the case but it crashes slightly more gracefully. There is room for improvement, such as if the user input the wrong file path, that biome is then not imported to the game. This is yet to be made however .